feat: added verifies and structure for the create job page
