+++ src/app/(authenticated-routes)/new-job/layout.tsx
@@ -1,0 +1,19 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Dev's Den - Abrir Vaga",
+  description: "Busque o talento desejado para a sua empresa."
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+
+  return (
+    <>
+        {children}
+    </>
+  );
+}
