--- src/app/jobs/new-job/styles.module.scss
@@ -1,2 +1,0 @@
-@import '@/app/globals.scss';
-
